code/online_payments/pyproject.toml
[tool.poetry]
name = "online_payments"
version = "0.1.0"
description = "Handle online payments for different providers like SimplePay"
authors = ["Rollet Developers <backend-team@rollet.app>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
